client/requirements.txt
gevent_socketio==0.3.6
Pillow==9.3.0
requests==2.28.1
